lexer, validation 모듈화
- lexer가 tokenize와 lex 수행
- parser가 run 수행 시 lex 후 validation check 수행.
- validation check module화
